+++ shared/utility/math/angle.hpp
@@ -114,8 +114,8 @@ namespace utility::math::angle {
 
     /*! @brief Solves for x in $a \sin(x) + b \cos(x) = c ; x \in [0,\pi]$
      */
-    inline float solveLinearTrigEquation(float a, float b, float c) {
-        float norm = std::sqrt(a * a + b * b);
+    inline double solveLinearTrigEquation(double a, double b, double c) {
+        double norm = std::sqrt(a * a + b * b);
         if (norm == 0) {
             throw std::domain_error(
                 "utility::math::angle::solveLinearTrigEquation - std::sqrt(a*a+b*b) == 0 => Any value for x is "
@@ -124,9 +124,9 @@ namespace utility::math::angle {
         }
 
         // Normalise equation
-        float a_ = a / norm;
-        float b_ = b / norm;
-        float c_ = c / norm;
+        double a_ = a / norm;
+        double b_ = b / norm;
+        double c_ = c / norm;
 
         if (std::fabs(c_) > 1) {
             throw std::domain_error("utility::math::angle::solveLinearTrigEquation - no solution |c_|>1");
@@ -134,7 +134,7 @@ namespace utility::math::angle {
 
         // Find alpha such that $\sin(\alpha) = a\_$ and $\cos(\alpha) = b\_$, which is possible because $a\_^2
         // + b\_^2 = 1$
-        float alpha = std::atan2(a_, b_);
+        double alpha = std::atan2(a_, b_);
 
         // Hence the equation becomes $\cos(\alpha)\cos(x)+\sin(\alpha)\sin(x) = cos(x-\alpha) = c\_$
         return alpha + acos_clamped(c_);
