+++ shared/utility/motion/splines/Footstep.hpp
@@ -21,15 +21,15 @@ namespace utility::motion::splines {
         /**
          * Initialization with lateral foot distance and support foot
          */
-        Footstep(const float& foot_distance, const bool& is_left_support_foot = true);
+        Footstep(const double& foot_distance, const bool& is_left_support_foot = true);
 
         /**
          * Set the lateral foot distance parameters
          */
-        constexpr void set_foot_distance(const float& foot_distance_) {
+        constexpr void set_foot_distance(const double& foot_distance_) {
             foot_distance = foot_distance_;
         }
-        [[nodiscard]] constexpr float getFootDistance() const {
+        [[nodiscard]] constexpr double getFootDistance() const {
             return foot_distance;
         }
 
@@ -51,29 +51,29 @@ namespace utility::motion::splines {
         /**
          * Starting position of current flying foot in support foot frame
          */
-        [[nodiscard]] const Eigen::Vector3f& get_last() const {
+        [[nodiscard]] const Eigen::Vector3d& get_last() const {
             return support_to_last;
         }
 
         /**
          * Target pose of current flying foot in support foot frame
          */
-        [[nodiscard]] const Eigen::Vector3f& get_next() const {
+        [[nodiscard]] const Eigen::Vector3d& get_next() const {
             return support_to_next;
         }
 
         /**
          * Returns the odometry change of the current step.
          */
-        [[nodiscard]] const Eigen::Vector3f& getOdom() const;
+        [[nodiscard]] const Eigen::Vector3d& getOdom() const;
 
         /**
          * Left and right, current or next pose of foot in world initial frame
          */
-        [[nodiscard]] const Eigen::Vector3f& getLeft() const {
+        [[nodiscard]] const Eigen::Vector3d& getLeft() const {
             return left_in_world;
         }
-        [[nodiscard]] const Eigen::Vector3f& getRight() const {
+        [[nodiscard]] const Eigen::Vector3d& getRight() const {
             return right_in_world;
         }
 
@@ -82,20 +82,20 @@ namespace utility::motion::splines {
          * The target foot pose diff is given with respect to next support foot pose (current flying foot
          * target).
          */
-        void stepFromSupport(const Eigen::Vector3f& diff);
+        void stepFromSupport(const Eigen::Vector3d& diff);
 
         /**
          * Set target pose of current support foot using diff orders.
          * Zero vector means in place walking.
          * Special handle of lateral and turn step to avoid foot collision.
          */
-        void step_from_orders(const Eigen::Vector3f& diff);
+        void step_from_orders(const Eigen::Vector3d& diff);
 
     private:
         /**
          * Static lateral distance between the feet
          */
-        float foot_distance;
+        double foot_distance;
 
         /**
          * Current support foot
@@ -106,15 +106,15 @@ namespace utility::motion::splines {
         /**
          * Pose diff [dx, dy, dtheta] from support foot to flying foot last and next position
          */
-        Eigen::Vector3f support_to_last = Eigen::Vector3f::Zero();
-        Eigen::Vector3f support_to_next = Eigen::Vector3f::Zero();
+        Eigen::Vector3d support_to_last = Eigen::Vector3d::Zero();
+        Eigen::Vector3d support_to_next = Eigen::Vector3d::Zero();
 
         /**
          * Pose integration of left and right foot in initial frame.
          * Set at "future" state taking into account next expected fot pose.
          */
-        Eigen::Vector3f left_in_world  = Eigen::Vector3f::Zero();
-        Eigen::Vector3f right_in_world = Eigen::Vector3f::Zero();
+        Eigen::Vector3d left_in_world  = Eigen::Vector3d::Zero();
+        Eigen::Vector3d right_in_world = Eigen::Vector3d::Zero();
     };
 
 }  // namespace utility::motion::splines
